Implement DiffableDataSource on FeedsViewController
